let ConfigurableScriptSource extends Wrapper
